+++ src/frontend/js/components/PaymentInterfaces/index.tsx
@@ -1,16 +1,17 @@
 import { lazy, Suspense, useEffect } from 'react';
 import * as Joanie from 'types/Joanie';
 import { handle } from 'utils/errors/handle';
-import { PaymentErrorMessageId, PaymentInterfaceProps } from 'components/PaymentInterfaces/types';
+import { PaymentErrorMessageId } from 'components/PaymentInterfaces/types';
 
 const LazyPayplugLightbox = lazy(() => import('./PayplugLightbox'));
 const LazyDummy = lazy(() => import('./Dummy'));
+const LazyLyraPopIn = lazy(() => import('./LyraPopIn'));
 
 /**
  * In charge of rendering the right payment interface according to the
  * `provider` property. Return null if the provider is not supported.
  */
-const PaymentInterface = (props: PaymentInterfaceProps) => {
+const PaymentInterface = (props: Joanie.PaymentInterfaceProps) => {
   const isNotImplementedProvider = !Object.values<string>(Joanie.PaymentProviders).includes(
     props.provider_name,
   );
@@ -31,6 +32,7 @@ const PaymentInterface = (props: PaymentInterfaceProps) => {
         <LazyPayplugLightbox {...props} />
       )}
       {props.provider_name === Joanie.PaymentProviders.DUMMY && <LazyDummy {...props} />}
+      {props.provider_name === Joanie.PaymentProviders.LYRA && <LazyLyraPopIn {...props} />}
     </Suspense>
   );
 };